refactor(cli): one owner for the published contract; flags are never silently dropped (#54)
Move the CLI's published contract into `symbiote_host::cli_schema`: the schema
identities, the `OPERATION_RISKS` classification that fixes the policy schema's
enum, the document builders, the published-document table, the canonical
serialization, and the fixture directory. The `symbiote` binary is now only the
interface. Three spellings of the fixtures path collapse to one function, the
duplicated dangerous-kind filter collapses to `dangerous_kinds()`, and the
builder-equals-fixture check runs in-process under `cargo test --lib` instead of
only through the binary.

Two behaviors the earlier passes left unsettled are now decided, each pinned by
a test that fails against the old behavior:

- `--write` is scoped to `schema`. With any other command it is a usage error
  (exit 1) that sends nothing, instead of being accepted and silently discarded.
- `schema` refuses `--json` rather than reinterpreting it. `--json` promises the
  versioned envelope, whose `result` must carry a `kind` the documents do not
  have; wrapping them would break the published envelope schema or invent a
  protocol result body. `symbiote schema --json` is a usage error that prints
  nothing.

The protected bytes are unchanged: plain `symbiote schema` stdout is exactly the
canonical committed fixtures, and `--write` still reproduces them byte-for-byte,
so the CI diff step stays green.

Verified from a clean `symbiote-host` rebuild (2298 files removed): fmt clean;
clippy --workspace --all-targets --locked -D warnings clean; cargo test -p
symbiote-host 105 passed / 0 failed. Mutating `command_id.minLength` in the
builder fails `cli_schema::tests::the_committed_fixtures_are_exactly_what_the_builders_emit`
and makes the CI step's exact commands (`schema --write` to a temp dir, then
`diff -ru`) exit 1.
